CKSVAR models

ThresholdStability.jl includes tools to convert censored and kinked structural vector autoregressive (CKSVAR) models into TVAR models.

Ox code to estimate CKSVAR models is available here.

ThresholdStability.CKSVAR_to_TAR — Function

CKSVAR_to_TAR(C, Cstar, βtilde, nlags)

Returns 1. a set of matrices corresponding to the CKSVAR model (with censored variable in levels) in TVAR form, and 2. a set encoding state-space constraints.

ThresholdStability.CKSVAR_to_companion — Function

CKSVAR_to_companion(C, Cstar, βtilde, nlags)

Converts CKSVAR model (with censored variable in levels) into companion form.

ThresholdStability.CKSVAR_to_companionFD — Function

CKSVAR_to_companionFD(F, Fstar, βtilde, nlags; diff=true)

Converts CKSVAR model estimated with the censored variable entering in first differences into companion form. The default setting diff=true returns the companion form with the censored variable entering in first differences. To retrieve the companion form with censored variable entering in levels, set diff=false.

Construction of automata

CKSVAR_to_TAR outputs a set of matrices generated in an order generated by the perms function (as well as a set encoding state space constraints.)

automaton_constructor provides a method to construct an automaton suitable for use with a CKSVAR model converted using CKSVAR_to_TAR, assuming this order. It is also suitable for any other set of matrices that is ordered in a similar way (see examples/AR2.jl).

ThresholdStability.perms — Function

perms(p)

Generates a vector of all sequences of length p made up of elements -1 or 1.

ThresholdStability.automaton_constructor — Function

automaton_constructor(Σ::AbstractVector{<:AbstractMatrix})

Generates an automaton tracking admissible transitions between states. Σ is a vector of matrices corresponding to a TAR model, assuming a default order generated by perms.

automaton_constructor(Σ::AbstractVector{<:AbstractMatrix}, seqlist)

Generates an automaton tracking admissible transitions between states given a custom list of regime sequences.

Automata can also be constructed manually using HybridSystems.jl.