Stability

Motivation

Let $\{\mathscr{X}_\sigma\}_{\sigma=1}^m$ be a partition of state space $\mathscr{X}\subseteq\mathbb{R}^p$ s.t. each $\mathscr{X}_\sigma$ takes the form of a convex polyhedron, ```math \mathscr{X}\sigma={x\in\mathbb{R}^p\mid E\sigma x\geq* 0, D\sigma x=0},

where the inequality ``\geq_*`` may be strict or weak.

The tools in [ThresholdStability.jl](https://github.com/samwycherley/ThresholdStability.jl) are intended to determine whether a discrete-time model of form ```math
\begin{align}x_{t+1}&=\Phi(x_t)x_t\\
&:=\sum_{\sigma=1}^mA_\sigma\mathbf{1}\{x_t\in\mathscr{X}_\sigma\}x_t\end{align}

is asymptotically stable, where $\mathbf{1}\{x\in\mathscr{X}_\sigma\}$ is an indicator function and $A_\sigma$ are given $p\times p$ matrices.

Sufficient conditions for asymptotic stability, in descending order of conservativeness, are that

the joint spectral radius (JSR) of the set $\Sigma=\{A_1,\dots,A_m\}$ is strictly less than 1;

the constrained joint spectral radius (CJSR) of the constrained switched linear system $(\Sigma, G)$, where $G$ is an automaton describing the possible transitions between states, is strictly less than 1;

the 'state-constrained joint spectral radius' (SCJSR) of the state-dependent switched linear system $(\Sigma, G, \{\mathscr{X}_\sigma\})$ is strictly less than 1.

In general, exact computation of the JSR, CJSR and SCJSR are NP-hard, but can be efficiently approximated via semidefinite programming or sum-of-squares programming (see e.g. Parillo and Jadabaie, 2008).

JSR and CJSR

We provide functions to compute tight upper bounds on the JSR and CJSR and to compute the spectral radius of individual matrices, each based on utilities from SwitchOnSafety.jl.

ThresholdStability.spectral_radius — Function

spectral_radius(A::AbstractMatrix)

Calculates the spectral radius of matrix A.

ThresholdStability.jsr — Function

jsr(Σ::AbstractVector{<:AbstractMatrix}; d = 2)

Calculates an upper bound on the joint spectral radius of a set of matrices Σ.

jsr(s::AbstractSwitchedSystem; d = 2)

Calculates an upper bound on the joint spectral radius for the switched linear system s.

Missing docstring.

Missing docstring for cjsr. Check Documenter's build log for details.

The jsr function applied to a state-dependent switched system (\Sigma, G, X) or a constrained switched system (\Sigma, G) ignores the state-space constraints and automaton and reports the JSR of \Sigma. Likewise, the cjsr function applied to a state-dependent switched system ignores the state-space constraints.

SCJSR

ThresholdStability.sosbound_γ — Function

sosbound_γ(s::StateDepDiscreteSwitchedLinearSystem, d)

Compute minimum value of $γ$ for which the sum-of-squares program is feasible.

Keywords

optimizer=...: choose an SDP solver. The default solver is CSDP.

tol=...: set tolerance. The default is tol=1e-5.

verbose=...: set level of detail reported during calculation process. The default is verbose=0.

initstep=...: set initial step size. The default value is initstep=1.1.

ThresholdStability.sosbound_gamma — Function

sosbound_gamma(s::StateDepDiscreteSwitchedLinearSystem, d; optimizer=nothing, tol=1e-5, verbose=0, initstep=1.1)

Alias for sosbound_γ.

ThresholdStability.sdpbound_γ — Function

sdpbound_γ(s::StateDepDiscreteSwitchedLinearSystem)

Compute minimum value of $\gamma$ for which the SDP is feasible.

For keywords, see sosbound_γ

ThresholdStability.sdpbound_gamma — Function

sdpbound_gamma(s::StateDepDiscreteSwitchedLinearSystem; optimizer=nothing, tol=1e-5, verbose=0, initstep=1.1)

Alias for bound_γ.

SCJSR calculation